GH Actions: add task to verify forbidden names test files are always up-to-date
This new job runs the script to regenerate the forbidden names sniff test case files and will fail the build if those are not up to date.

Note: Alternatively, I could set up the job to automatically update the files when needed and commit the updates to the current branch.