GH Actions: update matrix for PHP 8.1 release
Note: while "normally", a PHP 8.1 build would be added to the code coverage job and the PHP 8.0 build moved to the normal test job, this is not possible while we are still using PHPUnit < 9.5 for code coverage. Fixing that is still on the to do list.

Also: `--ignore-platform-reqs` is no longer needed for installing on PHP 8.1.