Composer: use PHPCSDevCS v 1.1.3
PHPCS 3.6.2 introduced a new sniff to the PSR12 standard, which enforces no blank line at the start of a class.

That sniff conflicts with the `Squiz.WhiteSpace.FunctionSpacing` sniff when the `spacingBeforeFirst` property is set to `1`.

The latest release of PHPCSDevCS removes the conflict and should allow the CS build to pass again.

Ref: https://github.com/PHPCSStandards/PHPCSDevCS/releases/tag/1.1.3