Classes/NewConstVisibility: account for PHP 8.1 final class constants
Prevent the sniff from not reporting on this issue when PHP 8.1 `final` class constants are used with the `final` keyword between the visibility modifier and the `const` keyword.

Includes unit tests.
Includes fixing a parse error in the test case file.
Includes touching up the documentation of the test file.