NewReturnTypeDeclarations: new check for non-standalone standalone types
The `void`, `mixed` and the new `never` types are "stand-alone" types, i.e types which are not allowed to be combined with other types, so they cannot be part of a type union, nor can they be nullable.

This is a cross-version compatibility issue as the name _could_ have been used, prior to the introduction of the type, as a class/interface name and could have been nullable or part of a union as the class/interface name.

In the `NewReturnTypeDeclarations` sniff, this was so far only addressed by detecting the use of the nullable operator in combination with the `mixed` type.

That particular check has now been refactored to be based on a list of types in the `$standAloneTypes` property and to check for both the nullability operator, as well as the use of the type in a type union.

This includes a minor BC-break in that the `PHPCompatibility.FunctionDeclarations.NewReturnTypeDeclarations.NullableMixed` error code has now been changed to `PHPCompatibility.FunctionDeclarations.NewReturnTypeDeclarations.NonStandaloneMixed` .

Includes unit tests and minor adjustments to the existing unit tests to accommodate.