FunctionNameRestrictions/NewMagicMethods: special case the magic __[un]serialize() methods
... in light of the PHP 8.1 deprecation of the `Serializable` interface.

For classes which implement the `Serializable` interface and need to support both PHP < 7.4 as well as PHP >= 8.1, the only way to avoid a deprecation warning is by implementing both the `Serializable` interface as well as the PHP 7.4 magic `__[un]serialize()` methods.

In that case, the sniff should stay silent as the code is cross-version compatible, as for PHP < 7.4, the `Serializable` interface methods will be used to serialize the object, while for PHP 7.4 and higher, the magic methods will be used.

Includes unit tests.

Fixes 1310