GH Actions: fix coverage not running on `develop`
Follow up on #1307 which broke the running of code coverage on the `develop` branch.

Apparently, an `if` conditional combined with a `needs`, MUST force the `if` to be read by calling `always()`, otherwise, if the `needs` was skipped (like we do for builds on `develop`), the condition will always return `false`.