Classes/NewConstructorPropertyPromotion: support PHP 8.1 readonly properties / intersection type
This commit adds tests to the sniff for PHP 8.1 `readonly` properties, as well as tests with PHP 8.1 intersection types and PHP 8.2 stand-alone true/false/null types.

While adding these tests, I discovered an oversight in how support was added for constructor property promotion to the PHPCS native `File::getMethodParameters()` and the PHPCSUtils `FunctionDeclarations::getParameters()` methods, as support for `readonly` properties without explicit visibility was not correctly handled.

A PR has been opened to fix this in PHPCS. In PHPCSUtils, this has been fixed in the `1.0.4` release, which is why the minimum supported PHPCSUtils version has been updated in the `composer.json` file.

As  the above means that the `'visibility_token'` index may not be available, the error will now be thrown on the `'token'` stack pointer, which points to the parameter name.