NewUnicodeEscapeSequence: make sniff more efficient
As things were, the sniff would check each and every text string token individually, while for multi-line text strings, this meant that the sniff would have to walk back to the first token in the text string for each line to ultimately be able to determine whether the text was using single or double quotes.

This minor refactor replaces the use of the `TextStrings::getCompleteTextString()` method with the new PHPCSUtils `TextStrings::getEndOfCompleteTextString()` method, examines multi-line text strings in one go and then skips over the rest of the text string tokens to prevent duplicate efforts.