Variables/ForbiddenThisUseContexts: add test with $this in enums
This is now automatically fixed after the update of PHPCSUtils to 1.0.0-alpha4.

This sniff uses a couple of token arrays which previously did not contain the `T_ENUM` token, but that token has been added to the relevant arrays as of PHPCSUtils 1.0.0-alpha4.

Report also contains similar unit test.

Fixes 1343