Composer: add PHPCSDevCS to the dependencies
PHPCSDevCS was previously not added to the `require-dev` dependencies as the minimum supported PHPCS version conflicted with the minimum supported PHPCS version of this package.

Now this is no longer the case, the package can be safely added to `require-dev` and work-arounds can be removed.

Note: includes declaring a `COMPOSER_ROOT_VERSION` environment variable in the test workflows as PHPCompatibility is a dependency of PHPCSDevCS and having the `COMPOSER_ROOT_VERSION` environment variable prevents a conflict due to the circular dependency.