Sniff: remove the `stripVariables()` method
... in favour of the significantly more comprehensive PHPCSUtils `TextStrings::stripEmbeds()` method.

This improves the handling of embedded variables in double quoted strings (and consequently fixes some bugs).

Includes adjusting all uses of the method in the repo.
Includes adding additional unit tests/adjusting existing tests to use more complex interpolated variables.
These tests would previously lead to false positives/false negatives. With the change-over to the PHPCSUtils version of the method, those false positives/negatives are fixed.

Includes an additional change in the `NewInterpolatedStringDereferencing` sniff where the use of the `TextStrings::getCompleteTextString()` method is replaced by the `TextStrings::getEndOfCompleteTextString()` to prevent having to duplicate some of the token walking to find the end of the text string within the sniff itself, as well as to safeguard a higher accuracy for the token walking (in case of tokenizer bugs, like I discovered while writing the PHPCSUtils method for stripping embeds - fixes for those bugs have, of course, been pulled to PHPCS and have been merged now).