PHP 8.0/8.1 | ForbiddenNames: detect use of new keywords when used as names
PHP 8.0 introduced `match` as a new reserved keyword and `mixed` as a new "other" reserved keyword.
PHP 8.1 introduced `readonly` as a new reserved keyword, `never` as an "other" reserved keyword and `enum` as a soft reserved keyword.

Note: `readonly` has an exception for when it is used as a function declaration name.

Includes regenerated test case files.

Refs:
* https://wiki.php.net/rfc/match_expression_v2
* https://wiki.php.net/rfc/mixed_type_v2#backward_incompatible_changes
* https://wiki.php.net/rfc/readonly_properties_v2
* https://wiki.php.net/rfc/enumerations
* https://wiki.php.net/rfc/noreturn_type#backwards_incompatible_changes
* https://github.com/php/php-src/pull/7468 (readonly exception in PHP 8.1)
* https://github.com/php/php-src/pull/9512 (readonly exception PHP 8.2 follow-up)