NewExecutionDirectives: `strict_types` does not type juggle
The `strict_types` execution directive allows for a value `0` or `1`. Declare statements do not type juggle, so a scalar value which could juggle to either of those types is not supported.

Also note that the representation of the token content in PHPCS is always a string, so the "valid values" to compare against need to be a text string, so we need to make sure that potential text string quotes are not stripped off.

Includes tests.