NewExecutionDirectives: improve message line precision
Previously, all messages would be thrown on the `T_DECLARE` token, while the actual directive token _could_ be on another line alltogether.

This changes the sniff to throw the messages on the actual execution directive instead of on the `declare` keyword.

Tested via adjusting an existing test.
Includes safeguarding the whitespace and comment tolerance in the sniff code (which is fine, but was insufficiently safeguarded).