OptionalToRequiredFunctionParameters: bug fix for `gmmktime()`
From https://github.com/php/php-src/blob/69888c3ff1f2301ead8e37b23ff8481d475e29d2/UPGRADING#L192-L194 :
> `mktime()` and `gmmktime()` now require at least one argument. time() can be
used to get the current timestamp.

This commit fixes a bug for the offset being wrong for the `gmmktime()` function (introduced in
https://github.com/PHPCompatibility/PHPCompatibility/commit/610e79ccb07f185cc47d9e1324029e88299c08c9 ).

Note: as the change for `gmmktime()` and `mktime()` is actually "require at least one argument", if we combine that with named parameters, we may need to change how we sniff this as currently, as we'd report when the `hour` param is missing, even when one of the other parameters would be passed.