ForbiddenNames: fix bug in enum special casing
In a few cases, the `enum` keyword will be tokenized as `T_STRING`, but should still be treated as the keyword.
The sniff was not selective enough about this though and would incorrectly treat a `enum` when used in a namespaced name as the `enum` keyword.

Fixed now.

Includes test.

Fixes 1474