NewKeywords: minor improvement
Prior to PHP 8.3, a `yield /*comment*/ from` sequence would be tokenized by PHP (and PHPCS) as `T_YIELD` + `T_COMMENT` + `T_STRING`.

Based on the current test failure on PHP 8.3, this will change in PHP 8.3 and the sequence will now tokenize as `T_YIELD_FROM`.

As the actual test is about `yield from` anyway, I've elected to make this work the same across all PHP versions and to always report this for the use of `yield from` instead of `yield`.

Includes updated test.