ParameterValues/HashAlgorithms*: refactor to use the AbstractFunctionCallParameterSniff
As the `NewHashAlgorithms` and the `RemovedHashAlgorithms` sniffs examine the contents of a function call parameter, it is more appropriate to base the sniffs on the `AbstractFunctionCallParameterSniff` so it can benefit from any "is this a function call ?" determination fixes made in that abstract.

This was previously not possible as the sniffs extended the `AbstractNewFeatureSniff` sniff, but after the refactor done in 1406, this is now possible.

Includes adjusting the `HashAlgorithmsTrait` which contained the pertinent "is this a function call ?" logic.

In the future, once a similar `AbstractFunctionCallParameterSniff` class is expected to be available via PHPCSUtils, this change will also allow us to switch over to that abstract more easily.

Note: The `HashAlgorithmsTrait::$hashAlgoFunctions` cannot be renamed to `$targetFunctions` as it would yield a fatal error:
`Fatal error: PHPCompatibility\AbstractFunctionCallParameterSniff and PHPCompatibility\Helpers\HashAlgorithmsTrait define the same property ($targetFunctions) in the composition of PHPCompatibility\Sniffs\ParameterValues\NewHashAlgorithmsSniff. However, the definition differs and is considered incompatible.`

For that reason, a `__construct()` method has been introduced to assign the value of the `HashAlgorithmsTrait::$hashAlgoFunctions` property to the `$targetFunctions` property, which is expected by the abstract sniff.