InitialValue/NewConstantArraysUsingDefine: refactor to use the AbstractFunctionCallParameterSniff
As the sniff examines the existence of specific function call parameters, it is more appropriate to base the sniff on the `AbstractFunctionCallParameterSniff` so it can benefit from any "is this a function call ?" determination fixes made in that abstract.

This straight away fixes some false positives. See the extra tests.

In the future, once a similar `AbstractFunctionCallParameterSniff` class is expected to be available via PHPCSUtils, this change will also allow us to switch over to that abstract more easily.