FunctionUse/NewFunctionParameters: refactor to use the AbstractFunctionCallParameterSniff
As the sniff examines the existence of specific function call parameters, it is more appropriate to base the sniff on the `AbstractFunctionCallParameterSniff` so it can benefit from any "is this a function call ?" determination fixes made in that abstract.

This was previously not possible as the sniff extended the `AbstractNewFeatureSniff` sniff, but after the refactor done in 1406, this is now possible.

In the future, once a similar `AbstractFunctionCallParameterSniff` class is expected to be available via PHPCSUtils, this change will also allow us to switch over to that abstract more easily.