FunctionUse/RemovedFunctionParameters: bug fix
Follow up on 1433 which added support for named parameter to the sniff.

Looks like one particular reference still presumed positional parameters and was not made compatible with named parameters.

Fixed now.

Note: no test included as at this time the `'callback'` array key appears to be unused, so there is nothing to test this with.