InitialValue/NewConstantArraysUsingDefine: prevent false negative
The code checking for an array token would only check the _first_ array-like token found and the sniff would end if that token was either not a (short) array or not at the correct nesting level.

However, a parameter can consist of multiple tokens, so _all_ array-like tokens in the parameter should be examined to prevent false negatives.

Fixed now.

Includes tests demonstrating the bug and safeguarding the fix.