InitialValue/NewConstantArraysUsingDefine: prevent false positives with PHP 8.1 "new in initializers"
Since PHP 8.1, initializers can contain objects.
As a side-effect of this, it is now allowed to initialize a constant with an arrow function or closure (as long as no parameters are declared), which would result in an object of the type `Closure`.

This sniff searches specifically for array tokens, but if such array tokens would be found _within_ a closure or arrow function definition, this would lead to a false positive.

This commit updates the sniff to skip over closure and arrow function declarations completely, thereby fixing the false positive.

Includes tests demonstrating the issue and safeguarding the fix.