Upgrade/LowPHP[CS]Sniff: minor efficiency tweak
The code in the sniffs only runs on the first file containing a PHP open tag, but a file may contain multiple open tags.

While the sniffs will only throw the error/warning once already, it could still trigger the sniffs a second time if multiple open tags are in a file.

This minor tweak will prevent that.