Keywords/NewKeywords: merge `NewArrowFunctionSniff`
The `NewArrowFunction` sniff was created prior to the decision to raise the minimum supported PHPCS version to 3.7.1.

As for older PHPCS version, a lot of custom logic was needed to recognize arrow functions, it made sense to have a separate sniff for this.

However, when support for PHPCS < 3.7.1 was dropped, the sniff was slimmed down to not contain any logic anymore, so we may as well merge the detection of arrow functions into the `NewKeywords` sniff.

This is **not** a BC break as the `NewArrowFunctionSniff` was new to PHPCompatibility 10.0.0 anyway.

Note: all tests from the original `NewArrowFunction` sniff have been moved over and merged into the `NewKeywords` test file.

Related to 808