UseDeclarations/NewUseConstFunction: bugfix - "as" keyword is case-insensitive
As things were, the `as` keyword was treated case-sensitively by the sniff, while the keyword is treated case-insensitively by PHP.

The sniff previously looked for the token content being `as`, as the tokenization could be different depending on the PHPCS version.

At this moment, that is no longer the case, so by switching to checking for the keyword based on the token _code_, the bug (false positives) gets fixed.

Includes a few adjusted tests to safeguard the issue.