Syntax/NewFunctionCallTrailingComma: bug fix - prevent false positives on function declarations
Trailing comma's in function declarations, as allowed per PHP 8.0, are not the concern of this sniff.

As things were, the sniff would throw false positives for trailing comma's in function declaration statements where the function was declared to return by reference.

Fixed now.

Includes simplifying the code a little as `T_CONST` and `T_USE` can never be found before a `T_STRING` token if there is a parenthesis opener after it.

Includes unit test.