Lists/NewListReferenceAssignment: tweak to improve performance
The `Lists::getAssignments()` function does a check on the short list related tokens to ensure the received token is a short list and it will throw an exception when this is not the case.

As the sniffs needs to listen for the `T_OPEN_SHORT_ARRAY` token (and sometimes the `T_OPEN_SQUARE_BRACKET` token as well), this means that in most examined files, an exception will be thrown in the majority of cases.

The creating of the exception + the `try-catch` have a (negative) performance impact.

This commit adds an early check on the received token to verify if it is a short list and bows out early (without the need for exception handling) when it's not.

While this does mean that there will be a duplicate call in the underlying code to `Lists::isShortList()`, the performance impact of this should be negligible as the result of function calls to `Lists::isShortList()` are cached.

Note: The `try-catch` around the call to `Lists::getAssignments()` is still needed, but will now only receive an exception when the `list` keyword is seen without closing parenthesis during live coding, which should be pretty rare.

Includes some additional unit tests to cover the unhappy path.

Related to #1477