Syntax/NewFunctionCallTrailingComma: bug fix - examine more tokens
Class instantiations with trailing comma's in the param list should also be detected by this sniff.

However, false positives on function declarations using reserved keywords should be prevented.
PHPCS normally (re-)tokenizes those function name in a declaration to `T_STRING`, but there seems to be a bug in this retokenization for the hierarchy keywords in combination with functions declared to return by reference.

Fixed now.

Includes tests.

Loosely related to 1489