Classes/RemovedOrphanedParent: bug fix for parent used as function name
Turns out PHPCS does not tokenize the `parent` keyword correctly as `T_STRING` when used as a function name in a function declaration which returns by reference, nor when that function name is subsequently called.
A fix has been pulled for this upstream: https://github.com/squizlabs/PHP_CodeSniffer/pull/3797

In the mean time, this commit fixes the issue for this sniff in a PHPCS cross-version compatible manner.

Includes unit tests.

Fixes 1489