Classes/RemovedOrphanedParent: detect use of `parent` in interfaces
Based on https://3v4l.org/7OT7b, it looks like support for the use of `parent` as a type declaration in interfaces has also been deprecated in PHP 7.4 and removed in PHP 8.0.

For interfaces, it doesn't matter whether the interface extends another interface or not.

This commit adjusts the sniff to detect this.

The sniff will throw a slightly adjusted error message and use a separate error code when the `parent` keyword is detected as used in an interface.

Includes unit tests.