DisableSniffMsg: change from trait to class with static method
As the method in this object is stand-alone anyway, there is no need for it to be a trait.

Includes renaming the method to a simpler name as with the method now being called statically, there is no need to repeat the information included in the class name in the method name.