DisableSniffMsg::create(): simplify the logic
As per the comment which was previously in the method, the delimiter lines were only needed as long as PHPCS < 3.3.1 was still supported.

Now support for those older PHPCS versions has been dropped, we can use new lines in the message text instead.

This also means that the `$phpcsFile` parameter is no longer needed, so the method signature has also been updated.

And as no `$phpcsFile` is needed, the test also no longer needs a `$phpcsFile` to be set up to run, so the `.inc` file can be removed and the parent class for the test class can be changed to the PHPUnit Polyfill native base class.