New `TokenGroup` helper for is*Number() methods
Move functionality which is only needed by select sniffs out of the base `Sniff` class.
This is a preparation step to be able to switch to using the upcoming PHPCSUtils sniff abstract classes in the future.

**Design notes:**
So far, most "helper" functions have been moved to `traits`. However, the `isNumber()` method included a call to the (non-static) `TestVersionTrait::supportsBelow()` method.

This is problematic as, while traits can use other traits, if the implementing class _also_ uses the same trait, PHP < 7.3 does not handle the (non-)conflict resolution correctly and this library still needs to be compatible with PHP 5.4.
See: https://3v4l.org/8c4kV

This is solvable by having conflict resolution in the implementing class(es), but that means this conflict resolution would need to be added in every class using the `TokenGroup`, which makes this unintuitive and error prone.
See: https://3v4l.org/nQbKV

So, to get around this, I've elected to make the `TokenGroup` a class with static methods instead and to add a new, required `$supportsPHP5` parameter to these functions.
That decouples the methods from needing access to the `TestVersionTrait` and makes the functions self-contained, while maintaining the same functionality.

Includes changing all relevant function calls to these methods and moving (and adjusting) the unit tests.