Classes/NewFinalClassConstants: rename the sniff
As `final` constants can now (also) be used in traits and enums, I propose to remove the `Class` from the sniff name. The sniff will still be in the `Classes` category.

As this is a new sniff in PHPCompatibility 10.0.0, this is not a BC-break.

Includes a small tweak to the error message to also remove the `class` keyword from the message.