Tests: improve performance
This is a next iteration to try and improve performance of the tests.

In a previous iteration, the processed `$phpcsFile` was already saved to an object local cache to improve performance.

However, the tests are still slow due to having to set up a new Config for each `testcase file - testVersion` combination, as well as having to tokenize the test case file again for each combination.

Cloning a pre-set-up `$phpcsFile` object is much faster than creating a new `Config`, `Ruleset` and `File` object.

This next iteration takes advantage of that by:
1. Splitting the set up and parsing (tokenizing) of the file from the processing.
2. Cloning the pre-set-up and parsed `$phpcsFile` object and changing the `testVersion` on the config property of the cloned object.
3. Only then processing the file for that `testVersion`.
4. And, of course, caching the results again.

On my local install this change brought the runtime of the complete test suite down from 15 minutes to 5 minutes, so I have a feeling this should help ;-)