Introduce two `InvalidTestVersion` exceptions
... and start using these in the `TestVersionTrait` instead of triggering a PHP warning.

For **end-users**, the net effect of this change is minimal.
* Previously, the warning would display as a PHPCS `Internal.Exception` on line 1 of the first file scanned.
    - The warning would not show if `error_reporting` did not include "warnings".
    - The warning would prevent the PHPCS run from finishing analyzing the file.
* Now, the (caught) exception will still display as a PHPCS `Internal.Exception` on line 1 of the first file scanned.
    - The exception can not be silenced anymore.
    - The exception will prevent the PHPCS run from finishing analyzing the file (same as before).

For **integrators**, the net effect of this change is an improvement.
* Previously, tools integrating PHPCompatibility could not "catch" the warning, only silence it.
* Now, the exception **_can_** be caught by tools which integrate PHPCompatibility.

On "our" side, it will make it easier to support PHPUnit 10.x at some point in the future.

Includes minor tweak to the language used in the error messages.
Includes perfunctory unit tests.
Includes adding the new directory to the test coverage configuration.