Change `TestVersionTrait` to `ScannedCode` class
As discussed during our meeting last month, this PR makes the following changes:
* Changes the `TestVersionTrait` trait to a `ScannedCode` class.
* Makes the methods `static`.
* Renames the `supportsAbove()` method to `shouldRunOnOrAbove()`.
* Renames the `supportsBelow()` method to `shouldRunOnOrBelow()`.

The aim of the change from `trait` to `class` and to `static` methods is to allow for these methods to be used in the helper methods in the  `TokenGroup` class as those methods otherwise would not have access to the `testVersion` check logic.

The aim of the renames is to allow the method calls to read more naturally and be more self-explanatory.

Includes updating all references to the trait (now class) and the methods.

Includes minor adjustments to the test code for these methods as the `getTestVersion()` method is `private`. The method was previously accessible as the `trait` was imported into the test class, but that's no longer the case and we do want to test the method via direct unit tests.