ScannedCode::getTestVersion(): refactor the code for improved performance
As things are, nearly every sniff calls either the `ScannedCode::shouldRunOnOrAbove()` or the `ScannedCode::shouldRunOnOrBelow()` method at least once, sometimes multiple times.

As things were, this would result in the `Helper::getConfigData('testVersion')` method call being executed hundreds, if not thousands of times during a PHPCS run, while the results would never change as the config data can not be changed during the run.

The refactor in this commit:
* Moves a couple of method local variables to static properties.
* Caches the results of the call to the `Helper::getConfigData('testVersion') method for better performance.
    As the `testVersion` can't be changed during a run, there will only ever be one result.
* Simplifies the `$arrTestVersions` variable / `$testVersions` property to a one-dimensional array.
    As the `testVersion` can't be changed during a run, there will only ever be one for the version array.
    Previously, this would be "cached" with the `testVersion` as the key. Now the result will be cached without that extra level in the array.

While this change should make PHPCS runs for PHPCompatibility a little faster, caching the results is problematic during the test runs where we _do_ change the `testVersion` on the fly.
For that reason, the local caching of the `testVersion` retrieved from the `Config`, as well as the caching of the determined `testVersions` is turned off when the `PHP_CODESNIFFER_IN_TESTS` constant is declared.

In addition to the above:
* The code in the `getTestVersion()` method has been re-ordered a little to lower the cyclomatic complexity.
* And the code in the `shouldRunOnOrAbove()` and `shouldRunOnOrBelow()` methods has received a little tweak which is possible now support for PHP < 5.4 has been dropped.