Remove the `PHPCompatibility.Upgrade.LowPHPCS` sniff
As per the decision in our meeting, we will as-of-now, update the minimum PHPCS version whenever needed to get access to Tokenizer improvements and will no longer create extensive work-arounds for missing Tokenizer backfills in individual sniffs/Helpers.

With that as new policy in mind, the `LowPHPCS` sniff can be removed as redundant.